# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Library API",
        "description": "API for managing books in library",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "operationId": "9017a3fa89d6636329067c858b27a6fa",
                "responses": {
                    "200": {
                        "description": "List of books"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create book",
                "operationId": "72f82e1cd6f08112323f87053277ca99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "publisher",
                                    "author",
                                    "genre",
                                    "publication_date",
                                    "words_count",
                                    "price_usd"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "publisher": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "genre": {
                                        "type": "string"
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "words_count": {
                                        "type": "integer"
                                    },
                                    "price_usd": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get book by ID",
                "operationId": "123856ec83e3466396b979a43eb1a15f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete book",
                "operationId": "eda12c526d0ccacaecea5f3d8f15d6df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Books"
                ],
                "summary": "Update book",
                "operationId": "3c5c1488774a14a2fa2e72f9ccee25d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "publisher": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "genre": {
                                        "type": "string"
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "words_count": {
                                        "type": "integer"
                                    },
                                    "price_usd": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Books"
        }
    ]
}